split out custom option classes to a new file
was just getting a bit big
